+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 Unreleased
 ------------
 - Added `bulk_update_with_history` utility function (gh-650)
+- Add default user and default change reason to `bulk_create_with_history` and `bulk_update_with_history`
 
 2.9.0 (2020-04-23)
 ------------------